% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_phenotype.R
\name{adjust_phenotype}
\alias{adjust_phenotype}
\title{Adjust Phenotype by Covariates}
\usage{
adjust_phenotype(pheno_file, covariate_file)
}
\arguments{
\item{pheno_file}{File path to phenotype.}

\item{covariate_file}{File path to covariates.}
}
\description{
This function adjusts the phenotype by the covariates.
}
