% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_memory_requirements.R
\name{approximate_memory_requirements}
\alias{approximate_memory_requirements}
\title{Estimate the memory requirements}
\usage{
approximate_memory_requirements(
  n_samples,
  n_snps,
  n_blocks,
  n_randvecs,
  chunksize
)
}
\arguments{
\item{n_samples}{Number of samples}

\item{n_snps}{Number of SNPs}

\item{n_blocks}{Number of genotype blocks}

\item{n_randvecs}{Number of random vectors}

\item{chunksize}{Number of focal SNPS per chunk}
}
\value{
Approximate memory requirement in giga bytes
}
\description{
Get an approximate estimation of the memory requirements
of the mme routine based on current input parameters.
}
\examples{
n_samples <- 1e5
n_snps <- 1e6
n_blocks <- 100
n_randvecs <- 100
chunksize <- 100
approximate_memory_requirements(n_samples,
                                n_snps,
                                n_blocks,
                                n_randvecs,
                                chunksize)

}
