% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fame.R
\name{fame}
\alias{fame}
\title{Fast Marginal Epistasis Test implementation}
\usage{
fame(
  plink_file,
  pheno_file,
  covariate_file,
  n_randvecs,
  focal_snp_index,
  n_blocks,
  rand_seed = -1
)
}
\arguments{
\item{plink_file}{File path for plink genotype dataset without *.bed extension.}

\item{pheno_file}{File path for plink *.pheno data.}

\item{covariate_file}{File path for covariates data.}

\item{n_randvecs}{Integer. Number of random vectors.}

\item{focal_snp_index}{Integer  of the focal SNP.}

\item{n_blocks}{Integer representing the number of blocks the SNPs will be read in.}

\item{rand_seed}{Integer to seed generation of random vectors. Only positive values are considered.}
}
\value{
A list of P values and PVEs
}
\description{
Fast Marginal Epistasis Test implementation
}
