% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fame_traits.R
\name{fame_traits}
\alias{fame_traits}
\title{Simulate traits from plink genotypes according to FAME paper}
\usage{
fame_traits(
  plink_file,
  trait_output_file,
  additive_variance_component,
  gxg_variance_component,
  focal_snp,
  gxg_indices,
  additive_indices
)
}
\arguments{
\item{plink_file}{Path to the plink data set without file extension}

\item{trait_output_file}{Path to the file that stores the simulated trait}

\item{additive_variance_component}{Float between 0 and 1 that quantifies the additive heritability of the trait}

\item{gxg_variance_component}{Variance component of the focal SNP}

\item{focal_snp}{Index of the focal SNP}

\item{gxg_indices}{Vector of SNP indices that are chosen to be epistatic with the focal SNP}

\item{additive_indices}{Vector of SNP indices that are chosen to have additive effects}
}
\value{
None
}
\description{
Simulate traits from plink genotypes according to FAME paper
}
