% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_gxg_groups.R
\name{get_groups}
\alias{get_groups}
\title{Split vector into groups}
\usage{
get_groups(gxg_indices)
}
\arguments{
\item{gxg_indices}{vector on SNP indices}
}
\value{
named list with indeces for the two groups
}
\description{
Split vector into groups
}
