% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdf5_file.R
\name{read_h5_file}
\alias{read_h5_file}
\title{Read the HDF5 file dataset}
\usage{
read_h5_file(filename, datasetName)
}
\arguments{
\item{filename}{File path to hdf5 file}

\item{datasetName}{name of the dataset to read}
}
\value{
Vector of indices
}
\description{
Read the HDF5 file dataset
}
