% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_api.R
\name{read_hdf5_dataset}
\alias{read_hdf5_dataset}
\title{Read Dataset from an HDF5 File}
\usage{
read_hdf5_dataset(file_name, dataset_name)
}
\arguments{
\item{file_name}{A character string specifying the path to the HDF5 file.}

\item{dataset_name}{A character string specifying the name of the dataset within the HDF5 file to read.}
}
\value{
The content of the dataset from the HDF5 file, typically in the form of an R object.
}
\description{
This function reads a dataset from an existing HDF5 file.
}
\examples{
\dontrun{
# Read a dataset named 'my_dataset' from an HDF5 file
data <- read_hdf5_dataset("example.h5", "my_dataset")
}
}
