% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_a_trait.R
\name{simulate_a_trait}
\alias{simulate_a_trait}
\title{Simulate a trait from plink genotypes}
\usage{
simulate_a_trait(
  plink_file,
  output_file,
  additive_heritability,
  gxg_heritability,
  additive_indices,
  gxg_indices_1,
  gxg_indices_2,
  log_level = "WARNING"
)
}
\arguments{
\item{plink_file}{Path to the plink data set without file extension}

\item{output_file}{Path to the file that stores the simulated trait}

\item{additive_heritability}{Float between 0 and 1 that quantifies the heritability of the trait}

\item{gxg_heritability}{Fraction of heritability due to additive only effects}

\item{gxg_indices_1}{Vector of SNP indices that are chosen to be epistatic in group 1}

\item{gxg_indices_2}{Vector of SNP indices that are chosen to be epistatic in group 2}

\item{log_level}{Log level.}

\item{n_causal}{Integer number of causal SNPs}
}
\value{
None
}
\description{
Simulate a trait from plink genotypes
}
