% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mask.R
\name{update_single_group}
\alias{update_single_group}
\title{Single group update}
\usage{
update_single_group(mask_file, gxg_group_1, gxg_group_2)
}
\arguments{
\item{mask_file}{Genotype mask}

\item{gxg_group_1}{snps in group 1}

\item{gxg_group_2}{snps in group 2}
}
\description{
Update the mask for a single group of snps.
}
