% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdf5_file.R
\name{write_h5_dataset}
\alias{write_h5_dataset}
\title{Write the HDF5 file dataset}
\usage{
write_h5_dataset(filename, dataset_name, new_data)
}
\arguments{
\item{filename}{File path to hdf5 file}

\item{dataset_name}{name of the dataset to read}

\item{new_data}{Vector to write to data set}
}
\value{
Writes to file
}
\description{
Write the HDF5 file dataset
}
