% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fame.R
\name{fame}
\alias{fame}
\title{Fast Marginal Epistasis Test implementation}
\usage{
fame(
  plink_file,
  pheno_file,
  annotation_file,
  gxg_bin,
  num_evec,
  snp_index,
  jack_number
)
}
\arguments{
\item{plink_file}{File path for plink genotype dataset without *.bed extension.}

\item{pheno_file}{File path for plink *.pheno data.}

\item{annotation_file}{File path for plink *.pheno data.}

\item{gxg_bin}{Integer.}

\item{num_evec}{Integer.}

\item{snp_index}{Integer.}

\item{jack_number}{Integer.}
}
\value{
A list of P values and PVEs
}
\description{
Fast Marginal Epistasis Test implementation
}
