% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mme.R
\name{mme}
\alias{mme}
\title{Fast Marginal Epistasis Test implementation}
\usage{
mme(
  plink_file,
  pheno_file,
  mask_file = NULL,
  grm_file = NULL,
  gxg_indices = NULL,
  chunksize = NULL,
  n_randvecs = 10,
  n_blocks = 100,
  n_threads = 1,
  rand_seed = -1,
  log_level = "WARNING"
)
}
\arguments{
\item{plink_file}{File path for plink genotype dataset without *.bed extension.}

\item{pheno_file}{File path for plink *.pheno data.}

\item{mask_file}{File path for gxg mask data.}

\item{grm_file}{File path for plink genotype dataset without *.bed extension that is only used for the GRM computation. If empty, plink_file will be used for the GRM computation.}

\item{gxg_indices}{List of indices for the focal variants.}

\item{chunksize}{Integer representing the number of SNPs analyzed with the same set of random vectors.}

\item{n_randvecs}{Integer. Number of random vectors.}

\item{n_blocks}{Integer representing the number of blocks the SNPs will be read in.}

\item{n_threads}{Integer representing the number of threads that are setup for OMP.}

\item{rand_seed}{Integer to seed generation of random vectors. Only positive values are considered.}

\item{log_level}{Log level.}
}
\value{
A list of P values and PVEs
}
\description{
Fast Marginal Epistasis Test implementation
}
