% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_traits.R
\name{simulate_traits}
\alias{simulate_traits}
\title{Simulate traits from plink genotypes}
\usage{
simulate_traits(
  plink_file,
  output_file,
  additive_heritability,
  gxg_heritability,
  n_causal,
  gxg_indices
)
}
\arguments{
\item{plink_file}{Path to the plink data set without file extension}

\item{output_file}{Path to the file that stores the simulated trait}

\item{additive_heritability}{Float between 0 and 1 that quantifies the heritability of the trait}

\item{gxg_heritability}{Fraction of heritability due to additive only effects}

\item{n_causal}{Integer number of causal SNPs}

\item{gxg_indices}{Vector of SNP indices that are chosen to be epistatic}
}
\value{
None
}
\description{
Simulate traits from plink genotypes
}
