% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_traits.R
\name{update_mask}
\alias{update_mask}
\title{Update mask}
\usage{
update_mask(mask_file, gxg_group_1, gxg_group_2)
}
\arguments{
\item{mask_file}{Genotype mask}

\item{gxg_group_1}{snps in group 1}

\item{gxg_group_2}{snps in group 2}
}
\description{
Update mask
}
